+++ CHANGELOG.md
@@ -7,6 +7,10 @@
   useful summary for people upgrading their application, not a replication
   of the commit log.
 
+## Unreleased
+
+* Remove options from public layout ([PR #5214](https://github.com/alphagov/govuk_publishing_components/pull/5214))
+
 ## 63.2.0
 
 * Allow summary card to support a block ([PR #5193](https://github.com/alphagov/govuk_publishing_components/pull/5193))